requirements.txt
# Django et REST Framework
Django==5.1.8
djangorestframework==3.16.0
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.5.3

# Celery et dépendances pour les tâches asynchrones
celery==5.5.0
django-celery-beat==2.7.0
redis==5.2.1
amqp==5.3.1
billiard==4.2.1
kombu==5.5.2
vine==5.1.0
python-crontab==3.2.0
cron-descriptor==1.4.5
django-timezone-field==7.1

# OpenAI pour le traitement du langage naturel
openai==1.70.0
pydantic==2.11.2
pydantic_core==2.33.1
annotated-types==0.7.0
typing-inspection==0.4.0
typing_extensions==4.13.1

# HTTP et requêtes
requests==2.32.3
urllib3==2.3.0
httpx==0.28.1
httpcore==1.0.7
h11==0.14.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
anyio==4.9.0
sniffio==1.3.1
jiter==0.9.0
distro==1.9.0

# Utilitaires
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pillow==11.1.0
six==1.17.0
tzdata==2025.2
tqdm==4.67.1

# Interface en ligne de commande
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
prompt_toolkit==3.0.50
wcwidth==0.2.13
